requirements.txt
amqp==5.1.1 ; python_version >= "3.9" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_full_version <= "3.11.2"
billiard==4.1.0 ; python_version >= "3.9" and python_version < "4.0"
celery==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
cron-descriptor==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
django-celery-beat==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
django-celery-results==2.5.1 ; python_version >= "3.9" and python_version < "4.0"
django-timezone-field==5.1 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.2 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework-simplejwt==5.2.2 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
kombu==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.9" and python_version < "4.0"
pyro4==4.82 ; python_version >= "3.9" and python_version < "4.0"
python-crontab==2.7.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.9" and python_version < "4.0"
serpent==1.41 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
